+++ src/components/views/auth/CaptchaForm.js
@@ -141,7 +141,7 @@ module.exports = React.createClass({
 
         return (
             <div ref="recaptchaContainer">
-                { _t("This Home Server would like to make sure you are not a robot") }
+                { _t("This homeserver would like to make sure you are not a robot.") }
                 <br />
                 <div id={DIV_ID}></div>
                 { error }